Integrating previous work into cegs portal
Previous work had been done in the https://github.com/ReddyLab/DHSSearch repository.
That was for some early experimenting but, as these things do, the code had settled
into a decent architecture.

This repository was created to get a better foundation to start with, using cookiecutter-django.
And now the old code (views, models, etc.) is being added.

Because cookiecutter-django has a bunch of pre-commit hooks for things like black, flake8, and
isort, some changes were made to the code. But it's basically the same code from DHSSearch.