Integrates DHS list with Genome browser
The list of DHSs updates as the genome browser scrolls.

Currently this is hacked in. The next step is to set up a real static
asset pipeline. Then I want to add a "callback" property to Genoverse
tracks so the same data from the request Genoverse makes to get the
DHSs can be used to build the table.