Adds a "gene" foreign key to GeneAssembly
The GeneAssembly<->Gene relationship is one-to-many -- a gene may be
in (have) multiple asssemblies, but each assembly only has one copy
of a gene.

Previously, I modeled this relationship as just a ManyToMany, but
lookups from the GeneAssembly (many) to the Gene (one) sides are
unnessarily complex this way. So I've added a foreign key to the
"many" side that points to the "one" so it's easy to find the Gene
associated with the GeneAssembly