Collapse Gene one-to-many migration
The original migration was applied to a database with data already in it,
so extra steps were taken to make sure the data could be updated to match
what the final schema was. Future databases will not already have data in
them when this migration is applied so we can get rid of the extra steps,
as they won't be applicable.