A script for loading GFF3 Gene annotation data
This script, when run in a django context*, will load the contents
of a gencode gff3 gene annotation file into the database.

While cleaning up the script so it's presentable and working on its speed (it's
pretty slow) I re-thought the GencodeGFF3 models and simplified them so instead
of having separate Annotation, Entry, and Attribute tables there is just the
Annotation table.

My initial impetus for separating Annotations and Entries was that a single annotation
(defined by the ID attribute of an entry) could have several entries. But looking at
the data and when that occurs (just for annotation types top_codon, three_prime_UTR,
start_codon, and five_prime_UTR) I decided that the extra complexity wasn't worth it
and combined Annotations and Entries.

With Annotation and Entries combined there was no longer any reason to have a whole separate
table for attributes, so that became a JSON field in the new Annotation table.

The loaddata make command in the Makefile was changed to run a script which calls
load_gencode_gff3_data.py.

The "local" database name changed from postgres (which is a default db that postgres comes with
and can't be dropped) to cegs_portal. This matches the prod DB name and also can easily be dropped and
re-created for testing purposes.

The migration files were re-created from scratch. The new DB model changes means that the data needs to
be reloaded. Might was well start the migration clean.

* e.g., python manage.py shell -c 'import load_gencode_gff3_data; load_gff3_data.run("filename", "GRCh38", "13")'