A script for loading wgCERES data
This script, when run in a django context*, will load the contents of a
a file produced from wgCERES assay data. This file should have a list of
DNase I Hypersensitive sites and regulatory effect information about each site.

The script will calculate, using the gene information already in the database,
the nearest gene to a given DHS by comparing the DHS midpoint to the TSS of
each gene.

Additionally it will add information about the experiment that produced the data
to the database.

The data this script was used on is in a shared CCGR Box folder called CEGS_portal
in the wgCERES directory.

* e.g., python manage.py shell -c 'from scripts import load_wgceres_data; load_wgceres_data.run("experiment.json")'