Load cCREs downloaded from SCREEN
The hg19 cCRE data was downloaded from SCREEN (https://screen-v10.wenglab.org). This
change includes a script to load the data into the DB. It also adds a new "File" model
that is used to indicate the source of the cCRE data.

cCREs are DHSs, so that's the table they were loaded into. The table was modified a bit -- "strand"
was removed because DHSs aren't on a strand. "source" and "screen_accession_id" were added, and
"cell_line" was changed to be nullable, as these cCREs are cell type-agnostic.