Steps 2&3 in replacing Gene/Transcript/Exon with generic Feature
This deletes all the references in the application to the old
Gene/Transcript/Exon models and updates queries to use Feature/FeatureAssembly.

There will be a step 3.5 to update the data loading scripts, but there are some
other, unrelated, changes in the scripts so I wanted that to be a seperate commit.