Fixes the feature rowspan on the search page
The rowspan was being set to 0, which worked fine in Safari but broke
the table in Firefox.

Now the rowspan should be set correctly.

I think this was a copy/paste issue from another table that did need
the ".all" because it was checking the length of a queryset or
manytomanymanager or something.