New query tokenizing algorithm
The old algorithm, which split the query on spaces and then checked
each part against the regexes had a downside that the regexes couldn't
match strings. so, e.g.,  "chr1: 1-10" couldn't be matched even though
we can easily write a regex for it.

The new algorithm doesn't have this issue.