Re-jiggers the Genoverse classes, fixes DHS Track state management
After removing the separate "DHSs with effects" track the class
structure was left more complicated than necessary. This
simplifies the class structure.

Also, it adds fixes a state management issue by not changing much, but
adding the line `this.browser.updateSharedState("dhs-data", data);` to
the ajax call that gets data from the backend. Unfortunately it's has
to be buried deep in the getData function, so that whole function gets
copied and slightly modified.