Use super() methods for get, get_json
This removes the get_template_prepare_data method in favor of
overriding get and using the super().get method.

This pattern wasn't possible while the data_handler function was
being passed around, but now that the data itself is being passed
around we can take advantage of the super() methods.