Chromosome visualization data format changes
The previous format tried to combine gene/ccre reg effect info within
a bucket into a "summary" that lost some information. This caused the
visualization to show ccres no corresponding genes (and visa-versa).

This new format keeps all the necessary information, but tries to do
so in a compact manner, both to keep file size down and to make processing
fast.

I also documented the new format, as it's more cryptic and less self-documenting
than the old format.