Make bucket size per-chromosome
This doesn't make the files much bigger, but does make using them
in the chromosome visualization easier.

With each chromosome setting its own bucket size zooming in on a
chromosome, by swapping a level1 chrom with a level2 chrom, takes no
special code to determine the bucket width (in px).