Add Feature fields to FeatureAssembly
This is step 1 of 4(?) in combining Feature and Feature assembly.

Step 1: Add Feature Fields to FeatureAssembly
Step 2: Migrate the Data
Step 3: Update code to use FeatureAssembly only
Step 4: Remove Feature from DB, finalize some FeatureAssembly changes

The purpose of this is to simplify the DB schema and gain some performance,
since there won't be so many joins to get all the necessary data.