Updates gencode loading script for featureassembly/feature merge
While trying to merging Feature and FeatureAssembly data (into just
FeatureAssembly) I found that a tacit assumption I made -- that the
FeatureAssemblies of a Feature would all have the same parent, was
wrong. To correct this I have to reload all the FeatureAssembly
data from scratch.

This leads to updating this script (Fortunately, the GencodeAnnotation/Region
data doesn't need to be changed).

Running it blind was pretty slow, so I also made so changes to speed it up.
1) FeatureAssemblies are saved in bulk
2) FeatureAssembly Parents are retrieved in bulk
3) GencodeAnnotation are not associated with a FeatureAssembly