Adds "facet_num_val" JSON field to FacetedModels
The current solution for "continuous" facets (facets that have numeric
values) is to treat them the same as discrete facets -- which means the facetvalue
table gets full of numbers. lots of numbers.

I never liked this, but originally I liked it more than this solution in this commit. But the
more I mull this over, the less I like it and am more convinced it will be unsustainable, as
the facetvalue table will just have too many rows.

I don't think keeping the numeric facet values in a JSON field is perfect either, but I think
it's more scalable, and possibly more performant, since it's fewer joins to get the data.