Reset DB migrations
A couple of hairy changes led me to believe that it's probably
easier to do a DB "reset". This means reloading all the data, but after updating
the scripts, this takes less than 90 minutes.

The two changes are
1) Combine FeatureAssembly and Feature
2) Move numeric facet values to a JSON column on the associated table.

Most of these model chanages were made and committed previously. This
commit squashes all the migrations to make future changes easier.