Renames, Combines level1 and level2 chromo viz scripts
At this point, a more accurate way to describe the visualization is
an "experiment coverage", so scripts have been renamed as such.

Most of the code in the level1 and level2 scripts was the, so this
finally combines the to two python scripts into one. Future changes
will be much easier.