Creates experiment_coverage endpoint
This endpoint will be used to get data for the experiment coverage
visualization. The client will pass the facet values in a request. The server
will load and filter the data, then respond with the values for each bucket.

Thus, the heavy lifting will be on the server, where we have more options for
making filtering fast.