Generate and read experiment coverage manifest file
In order to shorten "time to first paint" for the experiment visualization the
manifest file has been modified to include the data needed to display the visualization
with no filters selected. Additionally, the references for the different "levels" were
removed, as the client no longer needs them.

The experiment page has been update to read the new manifest format and display the
buckets from it. Some modifications have been made in anticipation of adding the server
calls to get the filtered data.

A new shell script has been added that will generate all necessary data files for the
visualization.