experiment_coverage endpoint can be POSTed and returns relevant data
The endpoint reads one of the relevant files for the experiment data
and just returns the keys. This is very basic, but from here we
can add the data processing and improve the data loading (currently, 211 mb
takes 3s to load, we'll need to do better if this is going to be "on demand"
filtering).