pickle and cache data, transform to simplified format
The data is now pickled instead of json, which means  the files are half the size
and take half the time to load. There may be even better gains that can be made in
the future?

To ameliorate the cost of loading the data from disc, the data is cached using python's
lru_cache from the functools module. This should reduce how often the data is loaded from
disc while a user is playing with the web UI.

A function that transforms from the disc format to the simplified display format has been
added. This will be used post-filtering.