Fix cont facet only bug
If no discrete facets were selected and a continuous facet was changed
the response from the backed was just the full set of data with no
filtering.

The fix was to add an new clause to the "shortcut" conditional to
only do the shortcut if we were also skipping the continuous facet
filtering.

After adding this clause I swapped the then/else code so the if-clause was
clearer.