Replace python exp. cov. filtering with rust implementation
The rust implmentation is several times faster both loading and
filtering data. There's probably plenty of room for improvement.

"exp_viz" is written in Rust as a python library using pyO3 and Maturin

https://github.com/pyo3/pyo3
https://maturin.rs/

maturin can be installed using pip, and handles building and installing
the library.

The algorithm used by the rust implementation is the same as the python
filter function.