Add DNAFeature Model type
This model is a combination of DNARegion and FeatureAssembly. The idea is for
it to replace both of those. This should simplify the codebase some. I also think
that it is more explicable than having two models that are almost exactly the same.

Future commits will have all the code changes needed to do the change-over