Increase size of "thing" fields in accession id
a "thing" field of just one letter limited how expressive the id could
be which would result in workarounds for clashes between "things". For
example: exon and experiment -- they both start with "ex". One could be
DCPE and the other DCPX, but which is which?

With four letters, one can be DCPEXPR and the other DCPEXON.