Refactored exp_viz filter to be more parallelizable
Some experiments show that a library like rayon, which implements
a par_iter function that creates an iterable that can run multiple
parts in parallel can, depending on "shape" of the data, give a big
performance boost -- the data can't be just all in one chromosome.

This commit doesn't add rayon, but it does ready the filtering function
to use rayon if we decided to.