Front-end for combining experiment visualizations
This adds UI for selecting experiments to combine the current experiment
with.

There is currently no error checking, and not every experiment has coverage, so this can
fail. Additionally, there's no checking that reference genomes are the same.

JS that is just used for the visualization was moved into the js/exp_viz directory.

Backend is forthcoming.