Fixes experiment list SQL
Removed "Tissue Types" from Experiment list because there are none.
Changed how "Cell Lines" is computed.

Previously, because of the way annotations work, an experiment with
multiple data files would be in the queryset of ExperimentSearch.all()
multiple times. These annotations were used to get the tissue types and cell
lines assocated with the experiment.

"Tissue Types" was removed because there is no tissue type information for the
experiments. "Cell Lines" was changed to be a string aggregation function. With these
changes the annotations on ExperimentSearch.all() will no longer cause the same experiment
to show up multiple times in the experiment list.