Standardizes and factors pageable data tables
Tables that might hold so much data they need to be paged are only
in a couple of places now, but it's a pretty common thing to need.

This factors the code out into two "partials" and standardizes the
form of the JSON data the API that feeds the tables.

This isn't a great solution -- there are two parts and there's a lot of
"bleed" in and out of the templates because they aren't really functions.
I do think it's better than copying and pasting the code though.