Featureloc query speedups
prefetching all the facet information can speed up featureloc with "regeffects"
api calls by 10x, with orders of magnitude fewer queries (depending on the query range)

Changes to the DNAFeature model made the select_related stuff moot, so it was deleted.

Also modified to not fetch facet info unless it was asked for.