Updates search string parser to get (possible) gene names
Changes the parsing algorithm some so that it "eats" its way through
the search string. This allows us to match genes last. There isn't
a good way to actually match gene names, so POSSIBLE_GENE_NAME_RE
is just a heuristic. Basically, if nothing else matches it might be
a gene name.

Also adds tests for the parser.