Only parse location OR terms, but not both, from search query
I don't think a location and a gene name (for isntance) makes sense
as a search term, so the parser will only return parse information
for one of those, based on what is first in the search query.

HOWEVER, it will also return some warning information so the user
can be told what part of their query was ignored.

Also ignored: multiple locations