Check facet boxes in and search using search query facets
Previously, a search URL like http://127.0.0.1:8000/search/results/?query=chr1:157909847-157959940%20hg19&facet=11
ignored the selected facet in the query.

The changes mean that a user can go to a search URL with a facet and get the expected results.

Additionally, the result table has been updated to be in line with the feature table output in tables.js
This will avoid visual "popping" as one table gets replaced with another.