Replace Cell Line and Tissue Type with Biosample
If a Cell Line is like a "class" a biosample is like an "object" of
that class.

Each biosample can have significant TBD metadata associated with that
differentiates it from other biosamples of the same cell line.

Some biosamples may also not belong to a well known/named cell line.
For instance, they may be patient samples.