Change Accession tracking to use DB
Initially, accession ID tracking was done in using files. This wasn't
the right decisioin. There was some manual updating of the files that
needed to happen with the experiment accession ids, which wasn't great.
Additionally, it just seemed strange to keep this metadata outside of
the database.

These changes fix that.