New tasks app
This app is for running I/O bound tasks in the background. Rather than
add Celery or something to the stack, my hope is that this is sufficient
for our humble website and its expected scale. If not, we can add a
purpose-build job queue, like celery, Q, or RQ in the future.

It has a simple interface of a single decorator "as_task" with an optional
description parameter. Using this decorator on a function will make the
function run as a background task rather than blocking the main thread.

Currently there is no way to call back to originating thread, so these
tasks should mainly involve the DB in some way.