Update dnafeature/reo access controls on background thread
Updating the access controls for a set of dnafeatures/reos when
an experiment is set private or archived can take a few moments. To
avoid locking the whole server up the work should be done in the
background.

There's some code in here for testing -- we don't want the new
signal handler going off during tests, so I added a way to disable
it.