Fix test_search tests
Some of the search tests were failing about half the time. I think
it's because search filters out exons and transcripts and the features
being created were of one of those types. The fix here is to specify
they types of the created features to _not_ be one of those.