Adds task endpoint for viewing tasks
This endpoint will be part of the response of other endpoints that
create long-running/asynchronous tasks, like the upcoming experiment
data request endpoint.

I don't think that tasks being available to any logged in user is a
security issue, so that's the level of security were offering around
this endpoint.