Task decorator changes take 2
The previous changes worked in the tests, but did not work in the "real world".
This commit kind of undoes those changes and opts instead for running the thread
as an "on commit" callback.

The issue is that DB changes in the main thread aren't actually written out to the DB
immediately -- "task.save()" isn't fully written out to the DB, because it's part of a
transaction. So if a separate thread starts and tries to write to a task that has been
passed in from the main thread the write may fail because the transaction in the first
thread hasn't been completed (the data hasn't been written to the DB) and the transaction
itself isn't shared between threads.

This attempt at a solution is to wait until the commit is made (using "transaction.on_commit()")
before starting the task. That way the ThreadTask will definitely exist in the DB before the
new task starts.