Model and ViewModel for requesting REO data by region
This system uses a 'Materialized View', which is a table computed
from the values in other tables. This table isn't automatically
updated, and would need to be manually be refreshed whenever new data
is added to the DB.

The "reo_sources_targets" materialized view speeds up the actual
query we need use to get the data by orders of magnitude, so I think
that the extra annoyances associated with materialized views are worth
it.