API endpoints for DL'ing select experiment data
Adds two new endpoints for generating and downloading select experiment
data.

The first endpoint is for querying the DB for experiment data.
Because it's hard to say how long the request will take, the work will be
done in the background. The response of the endpoint tells the user where
to get the data when it's done and give them a task url they query to see
if the query has been completed.

Once the query from the first endpoint is complete the second endpoint returns
the file with the data in it.