Fix data download expr. filter bug
In select cases the data downloads were including data for all
experiments, not just the selected experiment. This was due to
some missing parenthesis in the SQL WHERE clause when querying
for both source and target locations.

The old statement looks like

WHERE X OR Y AND Z

The new statment looks like

WHERE (X OR Y) AND Z.