Adds facet query, removes pages, updates tests
Since we're adding filtering to the experiment list page I want to
try removing paging.

This commit adds the query to get the facet values associated with
all experiments, and propogates those to the appropriate places.

It also removes pagination from the experiment list. We can add it back later
if we need to, but I suspect that with facet filtering it won't be necessary.

Tests have been updated as needed.