Improve the DNA Feature tab tests
The previous test only checked the first returned result to see
if a tab should be shown. This could result in a tab not being shown
even if it should be. For example, if the first result had no children
but the second result did then no "Children" tab would be shown.

The new tests check all returned results so this doesn't happen.