Add the huey task queue to the project
Because we want more robust task management and we want
task scheduling (like "do X once a week") we need a real
task queue rather than the hacked-together system we have now.

Adding huey and its configuration to the project is the first
step in replacing the old tasks with new ones.