Set Huey config up to run normal in DEBUG environments
Huey, when running in a DEBUG environment (like the QA environment)
runs in immediate mode, which doesn't use the task queue mechanisms
at all. This is convenient locally, but we want real tasks in the deployed
test environment.

This change lets us set immediate mode with an env variable so when
deploying it will not run in immediate mode, because we set HUEY_IMMEDIATE
to false.