Remove Tasks (Step 1) - delete model, update tasks
A real task queue (huey) has been added, so we can remove the old one.
This is step one, where we delete the model and replace the old "as_task"
decorator with huey's "db_task".

Step 2 will delete the tasks app altogether. We have to do in this in two steps so the migration to remove the table from the DB
happens.