Loading only de-dupes within a single experiment
Based on some email with a grad student this seems like the appropriate
amount of deduplication.

It can be helpful to see the per-experiment metadata for a given DNAFeature
and the way that makes the most sense is to have one copy of each DNAFeature
for each experiment, then deduplicate at a higher level, using a materialized
view or doing so in the UI somehome.

This level of deduplication will also help avoid issues of when to make a DNAFeature
private or acrchived because each DNAFeature will be owned by only one experiment
at most.