Remove lodash
the lodash version we used has some security issue. We only use
two functions, debounce and throttle, from the library, so an easy
solution is to just write those functions ourself. This has the added
bonus of shrinking the page load since we don't have to load an entire
library, just a few lines of code.