Fix issues with quickly clicking facets
Every time a facet is clicked on two htmx.ajax calls are made to update
parts of the UI to reflect the selected facets.

These calls need to be associated with a distinct elements so they don't
step on each others toes. Otherwise the later htmx call will cancel
the earlier one. Since the calls can be associated with different parts of
the UI, this can cause bugs where some of the UI doesn't get updated.