Renames TemplateJsonView to MultiResponseFormatView
The view allows responses in HTML, JSON, and TSV. The name doesn't match
what the code does in the least, so we're updating it.

Additionally, the json_format check doesn't really belong in the class
since it's only used in one set of classes. so json_format was moved
into those classes and genoverse support was removed from all the places
it wasn't really needed.

Finally, the default JSON renderer was changed to one that raises a
not-implemented error, which makes more sense as a response.