Modify logging levels
We're going to try using Duke's splunk instance to monitor logs. For this to work
we have to send logs to the container's stdout/stderr instead of logging to disk.
we also have to make sure everything relevant gets logged.

Additionally, we don't need django to mail us. We'll set up alerts in splunk.