Make sure pwd is in python import path list
Because of where the source_loc_extractor is called from in this script
the util module isn't available by default. The "root" portal directory,
where the util module is, isn't in the importpaths.

By setting PYTHONPATH=`pwd` the current directory, presumedly the
root portal directory, is added to the list of import paths.