Don't import duplicate cCREs
The source of cCRE truth is the list of cCREs for hg38. To get a
list of cCREs for hg19 we perform a "liftOver" from one genome to
the other.

This liftOver can result in "duplicate" cCREs, where multiple cCREs have
the same location. This doesn't make sense data-wise or biologically, so
we filter duplicates out and just use the first cCRE for a location.