Speed up genocode data loading
Speeds up gencode data loading by about 5x. This is because it uses the PostgreSQL COPY command (via psycopg2)
instead of the Django ORM to load the data. Instead of creating a lot of objects and bulk loading them the object data is
put into a StringIO object and loaded via copy_from which uses the PostgreSQL COPY under the hood.

See:
https://www.psycopg.org/docs/cursor.html#cursor.copy_from
https://www.postgresql.org/docs/current/sql-copy.html