Speeds up DCPEXR2 experiment data loading
Using the same COPY technique as the cCRE and Genocde data loaders.

The COPY technique was applied not just to loading the tested elements but
to associating them with cCREs. The necessitated some significant changes.

I also found, while working with the data, that it would be helpful to update the
search_dnafeature_id sequence when manually creating dnefeature ids so
using the auto-generated ids still works. That's why there's now a FeatureIDs
context manager isntead of just a next_feature_id function.