Send experiment combos as tree of set ops
In anticipation of adding arbitrary set operations (currently just intersection and union) to the experiments UI, this work models a set of operations on experiments as a tree of operations.
This allows us to perform any permutation on set operations on the experiments being explored.

This involves changes in both the front and back-end. The front end has to send a tree of operations and the backend has to accept a tree of operations, validate them, and perform them, before using the resulting
set of features in the filter.