Adjusts how associated cCREs are calculated
This makes two changes to associating cCREs with tested elements.

1) Generates the list of cCREs anew for each data load. Before we used
same list from SCREEN every time and so weren't finding associations with
"artificial" cCREs generated from previous data loads.

2) Changes how gRNA-based experiment look up/create cCREs. The initial cCRE/Tested Element
overlap list is generated using the gRNAs but when creating "artificial" cCREs we use
the location of the DHS the gRNA is associated with instead of the gRNA location.