Improves browser perf when downloading source features
The query run when genoverse requests source features wasn't specific enough and returned too many results.
Additionally the feature returned had too much information.

The first issue is being tackled by making more specific query properties. The second by not requesting the feature and all REOs and all targets.
That information was used by the menu. Now that information is request for a specific feature on demand when the menu is opened for that feature.