Ref genome selection works despite assembly parameter
When the user searches for a DNA feature, the feature with the most up-to-date reference genome (currently GRCh38) is shown. The backend code has been changed to accommodate this, but most changes affect other code, which creates more issues. This solution removes the need to deal with the backend altogether because there will be future changes to the way we search DNA features with assembly. The user will see all reference genomes in the dropdown, and if the reference genome doesn’t exist for the DNA feature, the user will be notified.