Separate experiment loading into frontend and backend
There is a new set of classes for loading Experiments, in the load_experiment.py
file. These represent a "backend" and interface that experiment data
being loaded into the database must match.

The [experiment info]_experiment.py files will now be "frontend" files that
read and extract the data needed to the format required for the backend
to load.

Experiments 8 and 9 have been modified to use the new backend. This also
means their "experiment metadata" files (not in source control) have been
updated.

All other experiments will also need to be updated.

The cCRE association has been removed because it also needs to be re-written
so we don't "shell out" to bedtools during data loading. It will be a decently
large task because we'll have to basically re-write bedtools intersect in
python.