Separate analysis loading into frontent and backend
There is a new set of classes for loading Analyses in the load_analysis.py file. These represent a "backend" and interface that analysis data being loaded into the database must match.

The [experiment info]_analysis.py files will now be "frontend" files that read and extract the data needed to the format required for the backend to load.

Experiments 8 and 9 have been modified to use the new backend. This also means their "analysis metadata" files (which are not in source control) have been updated.

Experiments 8 and 9 are the test bed, so once they can be completely loaded, including cCRE and analysis data, other experiments will be moved to the new loading system.