Add WHITENOISE_SKIP_COMPRESS_EXTENSIONS to settings
This tells whitenoise which file types not to compress during a
collectstatic run. We added three binary file types to this because
they can be large and just aren't very compressable. Trying to do so
slowed down the compresstatic step a lot.

This is a change from using --ignore flags because we don't actually
want these files ignored. We want whitenoise to serve them, we just
don't want the to be compressed.