Process and load data via http
This adds the ability to add experiment/analysis data from over the internet.

Users with the appropriate permissions can use the new uploads page to add an experiment or analysis
to the portal by submitting a metadata file, or the file's location. If the metadata's experiment
data or results data location is http or https the portal will download the data and load it into the database.

Much of this code is based on existing code from the scripts/data_loading directory.