Replace ReoSourceTarget* Views with Tables
The materialized views are slow to update, as they can't be updated
piecemeal. This makes them unscalable for loading data, as the more data
we have the longer it'll take to refresh them.

For this reason I'm replacing them with a table with the exact
same schema. When a new analysis is loaded run the ReoSourceTarget*.load_analysis()
method with the new analysis Id and the new analysis data will be added to the table.

This is much more scalable as, now, only the new data gets added, the whole table
doesn't have to get regenerated.