Auto-generate analysis coverage data
This adds a step to the data upload process that auto-generates the coverage
maps for the ideogram viz when an analysis is uploaded.

The Dockerfiles have been updated to ensure the cov_viz and cov_viz_manifest applications
are on the servers.

The genome data files were moved so they don't have to be copied into analysis directories
anymore. This is where they should have been the whole time.

I've also deleted a few files we don't use anymore in the data generation
directory. These used to be used for ccre stuff, but that's all done directly
in python code now.