Remove signal receivers, combine expr/an load tasks
Uploading experiments on test, because of the different
way huey is configured to run vs. locally where the code
was developed, didn't work.

These two changes, removing post-save signal receivers and
combining experiment and analysis load tasks, fix the two
problems.

I think there was a weird interaction between signals and huey that
broke things. There's a weak reference that was unable to be
serialized.

The experiment and analysis loading needs to be done serially instead
of in parallel so combining them into a single task ensures that
happens.