GET CSRF token for upload form
In the "local" environment CSRF protection is very lenient and the
previous code worked just fine.

In qa/production it's stricter. This change is to get a new CSRF token
every time we want to submit to the upload form, which mimics better
what happens in the browser (and works correctly).