Add pagination to _feature_sig_reg_effect partial
The data this partial displays could have hundreds of items in it. This isn't ideal
to show all at once, so it should be paginated.

This adds the pagination. In addition it modifies how big the table scroll view is calculated
so that there's always 90px underneath it for the pagination links to go.

The original partial was broken up into two partials that can be loaded separately so the modal aspect
gets loaded once and only the content is what gets paged. This keeps the "modal loading" animations
from firing every time the page changes.

I also generated the portion of the GET query that filters the data in the view to simplify
the templates.