Pass empty string instead of null expr desc to frontend
A "null" experiment description kind of broke some javascript in facet_filter.js that
expected a string, list, or DOM Element.

I could have just changed the Javascript, but it feels more "right" to send an empty string instead
of a null description.

I did add an assert to dom.js, so other cases will be easier to debug.