Cancel filtering fetches from experiment list page
Clicking around on the filter check boxes can cause multiple fetches to be "in flight"
at the same time. These may not necessarily return in the expected order (i.e. the most
resent "state" of the filter might finish before other states). This can result in an experiment
list that doesn't reflect the current state of the facet checkboxes.

This change adds a way to cancel getJson fetches and uses this to cancel previous
facet filter fetches when creating a new one so that only the most recent fetch finishes.