Changed viewport size in which dropdowns collapse
As the viewport size decreases, there comes a point where the dropdown items start to appear compressed. To prevent this, we have designed them to collapse at an earlier stage. For the sake of consistency, this collapse occurs approximately at the same viewport size as when the navbar menu collapses.