Fix search view model tests
These tests were in a file that wasn't discovered by pytest due to
the name not starting with "test_". We rename this file so pytest
picks it up.

Since it hadn't been run due to the aformentioned issue when "target only"
reg effects were removed from the reg_effects fixture the tests broke but
hadn't been fixed.