Adds facets to upload experiment tests
Without these facets these tests fail. The reason there's a check
to see if the facets already exist is because they are created when pytest
created the DB (i.e. `pytest --create-db` is run) but are deleted when
pytest is done. So the shouldn't be created when pytest is run with --create-db
but they should otherwise.