Support Uploads of full experiment data
It's annoying to rely on 3rd party hosting to get data into the portal
all the time (especially when, as of the data of these commits, the lab
doesn't have any 3rd party hosting).

This code adds support for uploading a tar.gz file with the metadata AND
data included.

The backend untars the file and then loads the file as usual, more or less.