Fix menu overflow issue
When a gene is clicked on and the menu shown, if it went off the page on the bottom
it caused scrollbars to appear instead of doing a "visible" overflow
of the container. This is beacuse of the overflow-x-auto, which makes overflow-y
auto when it is set to visible, for some reason.

We don't actually need overflow-x to be auto until the screen is too small
to hold the minimum width of the browser, so this change makes it overflow-x-visible
for small+ screens and overflow-x-auto for smaller than that. This means that,
usually, the menu won't cause scrollbars to appear.