Fix coverage color when only 1 item in bucket
When there's only one item in the bucket the coverage
color for that bucket is black, which is incorrect. It doesn't
match the color in the legend. This is because of a divide by 0
that happens because the "range" of the value is 0, as there's only one.

So we match the behavior of the legend in this case and return 0.5, which
makes the color the middle value of the color gradient for that bucket.