Only resize if the width actually changed
On iOS devices the "resize" event is dispatched when scrolling, for some reason.
In order to avoid constant "resizes" (that are to the same width, but still make network calls
to refetch the data) we keep track of the current "full width" and only resize if the new computed width
is different.